# refine murmurhash3_x64_128 for bloom_filter