1. refine `murmurhash3_x64_128` based on https://github.com/PaddlePaddle/benchmark/issues/151#issuecomment-517139284 

results on `pyramid_dnn` training on E5-2620 v3.  

|before this PR |after this PR|speedup
|---|---|---|
|185.18 s/epoch|182.4 s/epoch| 1.5%

2. do some code clean